third_party/WebKit/LayoutTests/external/wpt/mixed-content/optionally-blockable/no-opt-in/cross-origin-http/video-tag/top-level/keep-scheme-redirect/no-opt-in-allows.https-expected.txt

This is a testharness.js-based test.
FAIL opt_in_method: no-opt-in
                                 origin: cross-origin-http
                                 source_scheme: https
                                 context_nesting: top-level
                                 redirection: keep-scheme-redirect
                                 subresource: video-tag
                                 expectation: allowed assert_equals: The triggered event should match 'allowed'. expected "allowed" but got "blocked"
Harness: the test ran to completion.
